Move learning aid logic to data repository
